rename modal and update import
