+++ client/src/components/screens/auth/AuthScreen.js
@@ -1,10 +1,7 @@
 import { Box } from '@mui/material';
 import { useLocation } from 'react-router-dom';
 
-import Login from './Login';
-import Register from './Register';
-// import Copyright from '../../Copyright';
-import { Copyright } from '../../index';
+import { Login, Register, Copyright } from '../../index';
 
 function AuthScreen(){
     // use current web location to determine whether render sign-in/sign-up screen
