+++ client/src/components/screens/auth/Login.js
@@ -3,7 +3,7 @@ import { Box, Typography, TextField, Button } from "@mui/material";
 import { Link } from 'react-router-dom'
 
 import AuthContext from '../../../auth';
-import { MUIAccountErrorModal } from '../../index';
+import { AuthErrorModal } from '../../index';
 
 function Login(){
     const { auth } = useContext(AuthContext);
@@ -83,7 +83,7 @@ function Login(){
                     Login
                 </Button>
             </form>
-            <MUIAccountErrorModal/>
+            <AuthErrorModal/>
         </Box>
     )
 }
