+++ client/src/components/screens/auth/Register.js
@@ -3,7 +3,7 @@ import { Link } from 'react-router-dom';
 import { Box, Typography, TextField, Button } from '@mui/material';
 
 import AuthContext from '../../../auth';
-import { MUIAccountErrorModal } from '../../index';
+import { AuthErrorModal } from '../../index';
 
 function Register(){
     const { auth } = useContext(AuthContext); 
@@ -94,7 +94,7 @@ function Register(){
                     Create Account
                 </Button>
             </form>
-            <MUIAccountErrorModal/>
+            <AuthErrorModal/>
         </Box>
     )
 }
